+++ lightrag/api/run_with_gunicorn.py
@@ -8,7 +8,8 @@ import sys
 import signal
 import pipmaster as pm
 from lightrag.api.utils_api import display_splash_screen, check_env_file
-from lightrag.api.config import global_args, get_env_value
+from lightrag.api.config import global_args
+from lightrag.utils import get_env_value
 from lightrag.kg.shared_storage import initialize_share_data, finalize_share_data
 
 from lightrag.constants import (
