+++ lightrag/lightrag.py
@@ -24,7 +24,7 @@ from lightrag.constants import (
     DEFAULT_MAX_TOKEN_SUMMARY,
     DEFAULT_FORCE_LLM_SUMMARY_ON_MERGE,
 )
-from lightrag.api.config import get_env_value
+from lightrag.utils import get_env_value
 
 from lightrag.kg import (
     STORAGES,
